chore: organize better inscriptions util
