+++ src/utils/crypto.js
@@ -5,9 +5,6 @@ import { TESTNET, ASSUMED_TX_BYTES, BITCOIN_PRICE_API_URL, DEFAULT_DERIV_PATH } 
 import BIP32Factory from "bip32";
 import { ethers } from "ethers";
 import { ECPairFactory } from "ecpair";
-import axios from "axios";
-import { TURBO_API } from "@lib/constants";
-import LocalStorage, { LocalStorageKeys } from "@services/local-storage";
 
 bitcoin.initEccLib(ecc);
 const bip32 = BIP32Factory(ecc);
@@ -166,26 +163,3 @@ export function sortUtxos(utxos) {
     const sortedData = utxos.sort((a, b) => b.status.block_time - a.status.block_time);
     return sortedData.map((utxo) => ({ ...utxo, key: `${utxo.txid}:${utxo.vout}` }));
 }
-
-export const getInscriptionsForAddress = async (address) => {
-    const response = await axios.get(`${TURBO_API}/wallet/${address}/inscriptions`);
-    return response.data;
-};
-
-export const getOutpointFromCache = async (inscriptionId) => {
-    const key = `${LocalStorageKeys.INSCRIPTIONS_OUTPOINT}:${inscriptionId}`;
-    const cachedOutpoint = await LocalStorage.get(key);
-    if (cachedOutpoint) {
-        return cachedOutpoint;
-    }
-
-    const {
-        data: {
-            inscription: { outpoint },
-        },
-    } = await axios.get(`${TURBO_API}/inscription/${inscriptionId}/outpoint`);
-
-    await LocalStorage.set(key, outpoint);
-
-    return outpoint;
-};
